+++ index.html
@@ -28,6 +28,6 @@
   </div>
 </body>
 
-<script src="index.js"></script>
+<script type="module" src="main.js"></script>
 
 </html>
