+++ styles.css
@@ -149,7 +149,7 @@ body {
   justify-content: flex-end;
 }
 
-.add-form-row-asd {
+.add-form-row-box {
   display: flex;
   justify-content: space-between;
 }
